rest-api/pyproject.toml
[tool.poetry]
name = "unhcr-tss"
version = "0.4.0"
description = ""
license = "GPL-3.0-or-later"
authors = ["Pierre Guilbert <pierre.guilbert@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.9"
aiobotocore = "^2.4.2"
fastapi = "^0.97.0"
dynaconf = "^3.1.7"
uvicorn = "^0.20.0"
python-dateutil = "^2.8.2"
python-multipart = "^0.0.5"
requests = "^2.28.2"
aiofiles = "^22.1.0"
Pillow = "^9.4.0"
fastapi-mail = "^1.2.8"
azure-storage-blob = "^12.19.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pre-commit = "^2.16.0"
flake8 = "^5.0.4"
black = "^22.6.0"
uvicorn = "^0.20.0"
python-dotenv = "0.21.0"

[tool.poetry.group.dev.dependencies]
fastenv = "^0.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[[tool.mypy.overrides]]
module = "dynaconf"
ignore_missing_imports = true
